+++ TimetableManagementSystem/src/main/java/student_mgmt/Databases.java
@@ -11,16 +11,20 @@ import com.mongodb.client.MongoDatabase;
 public class Databases {
     private static MongoCollection<Document> TEACHER_SCHE_DATABASE;
     private static MongoCollection<Document> STUDENT_SCHE_DATABASE;
+    private static MongoCollection<Document> MESSAGE_DATABASE;
     private static String Schedule_databaseName = "Schedule";
+    private static String Message_databaseName = "Message";
 
     public static void config() {
         try {
             ConnectionString connString = new ConnectionString(
                     "mongodb+srv://pfy1:uol123@timetablemanagement.uq12hfp.mongodb.net/?retryWrites=true&w=majority");
             MongoClient mongoClient = MongoClients.create(connString);
-            MongoDatabase database = mongoClient.getDatabase(Schedule_databaseName);
-            TEACHER_SCHE_DATABASE = database.getCollection("TT_Sche");
-            STUDENT_SCHE_DATABASE = database.getCollection("Student_Sche");
+            MongoDatabase Schedule_database = mongoClient.getDatabase(Schedule_databaseName);
+            MongoDatabase Message_database = mongoClient.getDatabase(Message_databaseName);
+            TEACHER_SCHE_DATABASE = Schedule_database.getCollection("TT_Sche");
+            STUDENT_SCHE_DATABASE = Schedule_database.getCollection("Student_Sche");
+            MESSAGE_DATABASE = Message_database.getCollection("TT_Message");
             // System.out.println("Connected Successfully");
         } catch (Exception e) {
             System.err.println("Error connecting: " + e.getMessage());
@@ -39,4 +43,8 @@ public class Databases {
         return STUDENT_SCHE_DATABASE;
     }
 
+        public static MongoCollection<Document> MESSAGE_DATABASE() {
+        config();
+        return MESSAGE_DATABASE;
+    }
 }
